+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* [Smaller menu button when smaller timebar + new "Zoom reset" action + fix zoom out to maximum bug]()
+
 ## v2.13.0
 
 * Added script to run tests in CI/CD of GitHub. And then deploy the result to GH Pages, i.e. https://flower-platform.github.io/react-timeline-10000/